Add FIFO methodology for valuation
Calculate P&L and ROI based on FIFO methodology. This is a full revamp of the previous logic which was using "averages" and not fully correct. The logic for averages has also been refactor into the same helper functions. The code logic is prepared to add LIFO support, but the code has not been implemented.

Improvements for FIFO are still needed to improve logic on timestamp comparison, which is a requirement for LIFO and can be added then.

The application does not currently support changing between AVERAGE and FIFO and uses FIFO by default, but options to change the methodology "live" are possible given the current infrastructure.